fix JetWeights issue when reading from COOL (TrigConfStorage-03-00-81)
2015-02-11 Joerg Stelzer
	* fix JetWeights issue when reading from COOL
	* TrigConfStorage-03-00-81

2015-02-05 Joerg Stelzer
	* fix problem when loading L1 menu where CTP files are absent
	* TrigConfStorage-03-00-80

2015-01-28 Joerg Stelzer
	* SessionMgr: dont delete the ReplicaSorter (it leads to a crash)
	* TrigConfStorage-03-00-79

2015-01-26 Joerg Stelzer
	* CoolFix.cxx: new interactive executable to simplify fixing COOL folders at
	P1. Not complete yet, but solves ATR-9469
	* TrigConfStorage-03-00-78

2015-01-24 Joerg Stelzer
	* Making DB connection and retrial timeout configurable (ATR-9008)
	* TrigConfStorage-03-00-77
...
(Long ChangeLog diff - truncated)

+++ Trigger/TrigConfiguration/TrigConfStorage/src/TrigConfCoolL1PayloadConverters.cxx
@@ -552,7 +552,7 @@ TrigConfCoolL1PayloadConverters::readLvl1BGDesc(const coral::AttributeList & al)
       throw runtime_error("Read BLOB for ItemToBunchGroupMap of unexpected size!");
 
    const unsigned char* p = static_cast<const unsigned char*>(blob.startingAddress());
-   for (size_t i = 0; i < blob.size(); ++i,++p) {
+   for (long i = 0; i < blob.size(); ++i,++p) {
       unsigned char mask = (*p);
       codes.insert( make_pair(i,mask) );
    }
